use a batch file to execute a target's commands
We're now using a batch file (called command script internally) to execute
the commands of a target. This saves us hopefully from doing weird KDE
hacks, where the commands of a target must run in the same shell.

This is because of commands like this:
  cd foo
  do something

Before we've merged the commands into something like:
  cd foo && do something

Also, we don't have any issues with passing double quotes to cmd anymore.