fix regression wrt target names, surrounded by double quotes
test case:
QTDIR\tests\auto\qprocess\testProcessSpacesArgs\onespace.pro

This project was linked, although the exe was up to date.
Problem was, that the target name was "one space.exe" (with
double quotes), which is not a valid file name.
The double quotes must be removed, when using the target name
as file name.
I've added DescriptionBlock::fileName() for this.