make detection of the death of a child process more robust
The detection of the death of a child process entirely relied on the
IO completion port returning a failure for the read on stdout or stderr.
If no such failure happened, Process wouldn't detect the death of the
process.
Added a QWinEventNotifier that listens on the process handle to make
the detection more robust.

Change-Id: Ib5ae34950a9aa43edc5dc9109ff199687d3fd672
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>