Gracefully release job client in error case
When aborting the make process, because a target failed or such, it
could happen that the job client was still acquiring a token from the
job server. The user saw the disturbing message "JobClient destroyed
while still acquiring". Wait in this case until the acquisition is done
and release the job client afterwards.

Task-number: QTCREATORBUG-17182
Change-Id: I4a0c256d365a23887560d511b0f157a9f36a0df2
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>