Fix sub-jom detection for chocolatey shims
Do not walk the process tree up to find if some parent is called
"jom.exe" to find out whether we're a sub-jom, because this fails if a
non-jom which is called "jom.exe" calls the real "jom.exe". This is the
case for chocolatey shims. Instead, simply check for the existence of
the _JOMSRVKEY_ variable which is set by the top-level jom.

Task-number: QTCREATORBUG-15494
Change-Id: I90db7cf4b34cb6c3a03764cb5137f1f6e4af223d
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>